==> H1N1_analysis_models/01_Data_Cleaning_H1N1.ipynb <==
from sklearn.model_selection import train_test_split

import pandas as pd
import numpy as np
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
%matplotlib inline

X_train = pd.read_csv('data/training_set_features.csv', index_col='respondent_id')
y_train = pd.read_csv('data/training_set_labels.csv', index_col='respondent_id')

df = X_train.merge(y_train, on = 'respondent_id')

df.head()

df.isna().sum()

## Colums with Significant Missing Values

Will look at colums that have a significant amount of missing values and see what is to be done with them.

#### Health Insurance

df.health_insurance.value_counts()

len(df[(df.health_insurance == 0) | (df.health_insurance == 1)]) / len(df.health_insurance)

len(df[df.health_insurance == 0]) / len(df[(df.health_insurance == 0) | (df.health_insurance == 1)]) * 100

# Here I am taking those who responded with 0 and assuming I put 1 for missing values.
# The idea is to look at what percent of the data 0 would make up if I did this in a model
len(df[df.health_insurance == 0]) / len(df.health_insurance) * 100

# Here I am doing the same thing but instead assuming I put in 0 for 
# all missing values
len(df[df.health_insurance == 1]) / len(df.health_insurance) * 100

We only have $54.04\%$ of the values in health insurance. A significant amount of these values are missing. Looking up the amount of Americans with health insurance in 2018 I find that according to [ASPE](https://aspe.hhs.gov/reports/overview-uninsured-united-states-summary-2011-current-population-survey-0) about $16.1\%$ of Americans did not have health insurance. If I input 1, which has health insurance, the number of people who are 0, don't have health insurance, will go down from $12.03\%$ to $6.50\%$. This would be a worse representation of the number of Americans without health than is currently represented by the data. So, for this reason, I will drop this column entirely.

#### Employment Industry & Employment Occupation

df.employment_industry.value_counts()

df.employment_industry.isna().sum() / len(df.employment_industry)

df.employment_occupation.value_counts()

df.employment_occupation.isna().sum() / len(df.employment_occupation)

Both of these columns have about $50\%$ of their values missing. Since these columns were randomly assigned character strings based on an industry I cannot do more research about what the different industries are and how representative they are of the true population employment. So, for this reason, I will have to drop these two columns in their entirety. 

# Dropping the colums
df = df.drop(['health_insurance', 'employment_industry', 'employment_occupation'], axis= 1)
df.head()

# H1N1 Vaccine

Will Create a train, test split here for the H1N1 vaccine. Other notebooks will run models.

X = df.drop(['h1n1_vaccine', 'seasonal_vaccine'], axis=1)
y = df[['h1n1_vaccine']]

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=42)

# X_train.to_csv('data/X_train_h1n1.csv')
# X_test.to_csv('data/X_test_h1n1.csv')
# y_train.to_csv('data/y_train_h1n1.csv')
# y_test.to_csv('data/y_test_h1h1.csv')
